Change to idea IDE && Added menu first pages
Changed the IDE to the idea Webstorm. Changed menu and functionality. TODO: Look up the icon on the left of the mobile menu.
Added 3 photos with footer to test tailwind and next-ui.
TODO: See the blur thing. The function in next-ui is "isblured", collides with the webkit blur we usually use.